requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0
openai>=1.30.0
faiss-cpu>=1.8.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
sqlalchemy>=2.0.0
apscheduler>=3.10.0
requests>=2.31.0
tiktoken>=0.7.0
chromadb>=0.5.0
